viewer: relocate IfcHighlighter / IfcIsolator / CustomPostProcessor
Move the viewer-side three.js modules out of the mixed-purpose
src/Infrastructure/ into src/viewer/three/, alongside ThreeContext and
Picker. Pure file relocation — class names and surface unchanged.

  src/Infrastructure/IfcHighlighter.js     -> src/viewer/three/
  src/Infrastructure/IfcIsolator.js        -> src/viewer/three/
  src/Infrastructure/IfcIsolator.test.js   -> src/viewer/three/
  src/Infrastructure/CustomPostProcessor.js-> src/viewer/three/

Updated consumers: ShareViewer (the only production importer of all
three), HideToggleButton (IfcIsolator), and the
__mocks__/web-ifc-viewer.js jest.mock paths.

src/Infrastructure/ still holds GlbClipper (PR #1509 unifies/deletes it),
IfcElementsStyleManager / IfcCustomViewSettings / ViewRulesCompiler
(IFC view-rules — different concern), and small standalone helpers
(IfcColor, ColorHelperFunctions, PlaceMark, CutPlaneArrowHelper).

1147 tests pass; lint + typecheck clean.
