+++ src/viewer/ShareViewer.js
@@ -1,9 +1,9 @@
 import {IfcViewerAPI} from 'web-ifc-viewer'
-import IfcHighlighter from '../Infrastructure/IfcHighlighter'
-import IfcIsolator from '../Infrastructure/IfcIsolator'
 import IfcViewsManager from '../Infrastructure/IfcElementsStyleManager'
 import IfcCustomViewSettings from '../Infrastructure/IfcCustomViewSettings'
-import CustomPostProcessor from '../Infrastructure/CustomPostProcessor'
+import IfcHighlighter from './three/IfcHighlighter'
+import IfcIsolator from './three/IfcIsolator'
+import CustomPostProcessor from './three/CustomPostProcessor'
 import ThreeContext from './three/ThreeContext'
 import debug from '../utils/debug'
 import {areDefinedAndNotNull} from '../utils/assert'
